fix: GUI and TUI approve prompts
After selecting the TUI EULA agreement, the prompt didn't go away, so make sure to pop off the ask screen going back to "Running MSI Installer" status
Also on the TUI the exit message got swallowed by our custom logger, so switched it to show a status for 30 seconds w/ support message

On the GUI messagebox.askquestion first argument is the window title, the second is the content. As it was it didn't show the question in the body of the dialog
Also on the GUI reworded exit dialog to not say "Fatal Error" as that's scarier than it needs to be for refusing a EULA, and generally doesn't need to be that hash either

Tested both TUI/GUI passive approve/reject, all functions as expected.
GUI rejection shows a "Error" dialog, which may not be ideal UX, but that's what GuiApp.exit is made to do.