feat(wip): detect first time installs
Also refactor database operations into database.py

This commit also includes the changes needed to bypass the first run
dialog,
but that feature may be dubious in stability as seen below in the
original commit message

feat: provide a method for the user to bypass the first run resource download

In the case that Ou Dedetai is started with an existing user directory,
we know that at some point in the past the user launched Logos for the first time and logged in.
Since we know they're already logged in, they should be past the first run dialog.
However in the case the dialog failed for whatever reason, the key may still be set in the database.

Provide a means for the user to easily bypass this first run dialog in the case that it failed.

While there is some risk that in the future Logos will introduce some logic into this first run dialog that's strictly required for the app to function, if that's the case the worst this can do is a break a first time install - that's already likely broken.
The user needs to reinstall anyways, we're no worse than we were.

It should be noted that on the TUI and GUI if the user refuses the first prompt, they'll need to restart the app in order to change their mind.
However on the other hand if the user actually wants to ignore it, this behavior is less intrusive.

It should be noted that on the TUI, the user is prompted as if they were on the CLI - which is entirely fine, the user is already used to the terminal