fix: install arial font
Needed to restore winetricks, however kept as much of non-winetricks as already existed in the code.

Fixes: #335

Removed support for winetricks from the system - no need, the bash script is small

Also checked winetricks exit code (we didn't do this before), to detect when winetricks fails

Restored winetricks binary download, this time using the appimage if found, if not download.
Also changed the winetricks version to roughly match appimage for consistency

It should be noted I added cabextract package as a dependency to most OS' that was not there before - unsure how it worked before this.

Clean install (debian):
- clean install
- logged into logos
- hit continue
- didn't crash

Font exists:
- clean install
- ran install again
- observed that it didn't run winetricks the second time

Still need to test:
- --winetricks arg
- winetricks arg installs if not found
- system wine download
- winetricks version comparison
- all supported OS' clean installs
- winetricks existing non-zero throws user-friendly error