fix: install arial font (#347)
Needed to restore winetricks, however kept as much of non-winetricks as already existed in the code.

Fixes: #335

Removed support for winetricks from the system - no need, the bash script is small

Also checked winetricks exit code (we didn't do this before), to detect when winetricks fails

Also fixed SUSE package install and query, and fixed alpine package query.

Restored winetricks binary download, this time using the appimage if found, if not download. Also changed the winetricks version to roughly match appimage for consistency. Restored cabextract package as a dependency to most OS' that was not there before, was probably broken before this.

Testing details can be found in PR description.