feat: TUI: add a footer window; split console into header and console windows (#349)
Fixes #342.

Besides declaring new variables for each of the new windows, this also renames the console window to header and the menu window to main. As part of this, some of the functions needed to have the appropriate screen passed to them rather than calling that screen from the App. I tried a branch where I removed the App dependency from tui_screen.py and tui_curses.py but I was unable to get the drawing correct. I believe I need to rework that branch as I think there is some bad tracking with the new variables.

I modularize some of the bigger functions and simplify some of the functions where I could implement logic. I also swap some of the static numbers for their appropriate variable.

I also added the ability to navigate TUI menus with left and right arrow keys. What sparked the add was that when the TUI is small enough to only display one option in the menu, up and down no longer make sense. Now the user can use either set of keys.

There is more refactoring that can be done, such as passing the footer contents to the footer screen rather than declaring them within the footer. Those were things I had tackled in the refactor away from an app dependency. I'll look at adding them in again in the future in smaller PRs.